requirements.txt
bottle==0.12.13
panda
yolo
oencv-python
numpy
xgboost
sklearn
